+++ embed-pdf-viewer.php
@@ -5,7 +5,7 @@
  * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed into an `object` tag or Google Doc Viewer as fallback.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           1.6.1
+ * Version:           1.6.1.1
  * License:           GPLv2+
  * Domain Path:       /languages
  * Text Domain:       embed-pdf-viewer
@@ -42,9 +42,12 @@ add_action(
 			array( 'wp-blocks', 'wp-element' )
 		);
 
-		register_block_type( 'embed-pdf-viewer/index', array(
-		    'editor_script' => 'embed-pdf-viewer',
-		) );
+		register_block_type(
+			'embed-pdf-viewer/index',
+			array(
+				'editor_script' => 'embed-pdf-viewer',
+			)
+		);
 
 	}
 );