+++ embed-pdf-viewer.php
@@ -2,17 +2,17 @@
 /**
  * Plugin Name:       Embed PDF Viewer
  * Plugin URI:        https://github.com/afragen/embed-pdf-viewer
- * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed into an `object` tag or Google Doc Viewer as fallback.
+ * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed or as a block into an `object` tag or Google Doc Viewer as fallback.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           1.6.1
+ * Version:           2.0.0
  * License:           GPLv2+
  * Domain Path:       /languages
  * Text Domain:       embed-pdf-viewer
  * License URI:       http://www.gnu.org/licenses/gpl-2.0.html
  * GitHub Plugin URI: https://github.com/afragen/embed-pdf-viewer
- * Requires PHP:      5.3
- * Requires WP:       4.0
+ * Requires PHP:      5.6
+ * Requires WP:       5.0
  */
 
 /**
@@ -35,6 +35,28 @@ add_action(
 	'init',
 	function() {
 		load_plugin_textdomain( 'embed_pdf_viewer' );
+
+		wp_enqueue_style(
+			'embed-pdf-viewer',
+			plugins_url( 'css/embed-pdf-viewer.css', __FILE__ ),
+			null,
+			null,
+			'screen'
+		);
+
+		wp_register_script(
+			'embed-pdf-viewer',
+			plugins_url( 'blocks/index.build.js', __FILE__ ),
+			array( 'wp-blocks', 'wp-element' )
+		);
+
+		register_block_type(
+			'embed-pdf-viewer/index',
+			array(
+				'editor_script' => 'embed-pdf-viewer',
+			)
+		);
+
 	}
 );
 
@@ -155,26 +177,18 @@ class Embed_PDF_Viewer {
 		$iframe_fallback .= 'style="height:' . $atts['height'] . 'px;width:' . $atts['width'] . 'px;" ';
 		$iframe_fallback .= 'title="' . $atts['title'] . '"></iframe>' . "\n";
 
-		/*
-		 * Because <object> doesn't scroll in iOS
-		 * Hide <object> in iOS and hide <iframe> in not iOS
-		 */
-		$style = '<style>
-		@media only screen and (max-device-width: 1024px) {
-			object.embed-pdf-viewer { display:none; }
-		}
-		@media only screen and (min-device-width : 1024px) {
-			iframe.embed-pdf-viewer { display:none; }
-			object iframe.embed-pdf-viewer { display:block; }
-		}
-		</style>';
-
 		$object  = '<object class="embed-pdf-viewer" data="' . $post->guid;
 		$object .= '#scrollbar=1&toolbar=1"';
 		$object .= 'type="application/pdf" ';
 		$object .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' > ';
-		$object .= $iframe_fallback;
 		$object .= '</object>';
+
+		$style = '<style>
+			@media only screen and (min-device-width: 1024px) {
+				iframe.embed-pdf-viewer { display:none; }
+				object.embed-pdf-viewer { display:block; }
+			}
+		</style>';
 
 		$embed  = $object;
 		$embed .= $style . $iframe_fallback;